requirements.txt
scrapy
SQLAlchemy
psycopg2-binary
instagrapi
Pillow
kafka-python3
kafka-python
pandas
django
djangorestframework
django-celery-beat
django-softdelete
celery
redis
numpy
pytz
psycopg2-binary
pyyaml
Jinja2